added post using AJAX

+++ controllers/posts_controller.js
@@ -3,14 +3,25 @@ const Comment=require('../models/comment');
 
 module.exports.create= async function(req,res){
     try{
-        await Post.create({
+       let post= await Post.create({
             content:req.body.content,  //req.body.content is from 'form' in home.ejs
             user:req.user._id
         });
+        if(req.xhr){
+            return res.status(200).json({
+                'data':{
+                    'post':post
+                },
+                'message':'post created!'
+            })
+        }
+        req.flash('success','post published!');
         return res.redirect('back');
     } catch(err) {
-        console.log(`Error: ${err}`);
-        return;
+        // console.log(`Error: ${err}`);
+        req.flash('error',err);
+        return res.redirect('back');
+
     }
         
 }
@@ -19,19 +30,30 @@ module.exports.destroy=async function(req,res){
     try{
         //req.params['id']==req.params.id
     let post=await Post.findById(req.params['id']);
+    if(req.xhr){
+        return res.status(200).json({
+            data:{
+                post_id:req.params.id
+            },
+            message:"post removed!"
+        })
+     }
     // .id means converting the object id into string
       if(post.user==req.user.id){
           post.remove();
-
          await Comment.deleteMany({post:req.params.id});
+         
+         req.flash('success','post deleted');
          return res.redirect('back');
       }
       else{
+        req.flash('success','you cant delete this post');
           return res.redirect('back');
       }
     }
     catch(err){
-        console.log(`Error: ${err}`);
-        return;
+        // console.log(`Error: ${err}`);
+        req.flash('error',err);
+        return res.redirect('back');
     }
 }
